+++ R/automation.R
@@ -43,11 +43,6 @@ setMethod("timestamps", "Script", function(x) {
 
 #' @rdname describe-catalog
 #' @export
-setMethod("scriptBody", "ScriptCatalog", function(x) {
-    return(getIndexSlot(x, "body"))
-})
-#' @rdname describe-catalog
-#' @export
 setMethod("scriptBody", "Script", function(x) {
     return(x@body$body)
 })
@@ -91,24 +86,28 @@ setMethod("scriptSavepoint", "Script", function(x) {
 
 #' Run a crunch automation script
 #'
-#' Crunch automation is a custom scripting syntax that allows you to
-#' concisely describe the metadata of your data when importing. The
-#' syntax is described [in the crunch API documentation](
-#' https://help.crunch.io/hc/en-us/categories/360004356012-Crunch-Automation)
+#' Crunch Automation is a custom scripting language that allows you to
+#' execute common Crunch commands. The syntax is described in the
+#' [Crunch API documentation](
+#' https://help.crunch.io/hc/en-us/categories/360004356012-Crunch-Automation).
 #'
-#' @param dataset A crunch dataset
-#' @param script A path to a text file with crunch automation syntax
-#' or a string the syntax loaded in R.
+#' If a character vector with length bigger than 1 is passed to `script`,
+#' it's converted to a string by concatenating its elements together using
+#' line breaks.
+#'
+#' @param x A crunch dataset or project folder (for backwards compatibility,
+#' `dataset` is also accepted)
+#' @param script A path to a text file containing a Crunch Automation script
+#' or a character vector of length 1 or more with Crunch Automation commands (see `Details`)
 #' @param is_file The default guesses whether a file or string was
 #' used in the `script` argument, but you can override the heuristics
 #' by specifying `TRUE` for a file, and `FALSE` for a string.
 #' @param encoding Optional encoding to convert **from**, defaults to UTF-8.
 #' The API accepts only UTF-8, so all text will be converted to UTF-8 before
 #' being sent to the server.
-#' @param ... Additional options, such as `dry_run = TRUE` passed on
-#' to the API
-#'
-#' @return For `runCrunchAutomation()`: an updated dataset (invisibly),
+#' @param ... Additional options, such as `dry_run = TRUE`, passed on
+#' to the API if x is a dataset (if x is a project folder, an error is thrown)
+#' @return For `runCrunchAutomation()`: an updated dataset/project folder (invisibly),
 #' For `showScriptErrors()`, when run after a failure, a list with two items:
 #' `script`: that contains the script string sent to the server and `errors` which is a
 #' `data.frame` with details about the errors sent from the server.
@@ -128,45 +127,109 @@ setMethod("scriptSavepoint", "Script", function(x) {
 #' # But more details are available with function:
 #' showScriptErrors()
 #'
-#' # After a successful run, can look at scripts
+#' # After a successful run, can look at scripts:
 #' scripts(ds)
 #'
+#' # Run Crunch Automation on a folder:
+#' my_folder <- cd(projects(), "folder1")
+#' runCrunchAutomation(my_folder, 'CREATE FOLDER "folder2";')
 #' }
-#' @export
 #' @seealso [`automation-undo`] & [`script-catalog`]
+#' @export
 runCrunchAutomation <- function(
-    dataset,
+    x,
     script,
     is_file = string_is_file_like(script),
     encoding = "UTF-8",
-    ...
-) {
-    reset_automation_error_env()
-    stopifnot(is.dataset(dataset))
-    stopifnot(is.character(script))
-
-    if (is_file) {
-        automation_error_env$file <- script
-        script <- readLines(script, encoding = encoding, warn = FALSE)
-    } else {
-        automation_error_env$file <- NULL
+    ...) {
+    # a previous version of this function had `dataset` as its first argument
+    # because the function is now broader, the argument is called `x`
+    # the following block is for backwards compatibility
+    args <- as.list(sys.call())
+    args[[1]] <- NULL
+    if ("dataset" %in% names(args)) {
+        warning(
+            "The argument `dataset` has been renamed to `x` because the function ",
+            "now supports more generic operations. ",
+            "While using `dataset` will continue to work, it will emit this warning."
+        )
+        target <- list(x = args$dataset)
+        args$dataset <- NULL
+        out <- do.call(
+            "runCrunchAutomation", c(target, args),
+            quote = FALSE, envir = parent.frame()
+        )
+        return(out)
     }
-    if (length(script) != 1) script <- paste(script, collapse = "\n")
 
-    body <- list(body = script, ...)
+    filename <- NULL
+    if (is_file) {
+        filename <- script
+        script <- readLines(script, encoding = encoding, warn = FALSE)
+    }
+    script <- paste(script, collapse = "\n")
+    reset_automation_error_env()
+    automation_error_env$file <- filename
     automation_error_env$last_attempted_script <- script
 
+    sendCrunchAutomationScript(x = x, script = script, ...)
+
+    # Provide feedback for dry_run success so that user is confident it was successful
+    if (isTRUE(list(...)$dry_run)) {
+        message("Script dry run was successful")
+        return(invisible(x))
+    }
+
+    invisible(refresh(x))
+}
+
+setMethod("sendCrunchAutomationScript", "CrunchDataset", function(x,
+                                                                  script,
+                                                                  ...) {
     crPOST(
-        shojiURL(dataset, "catalogs", "scripts"),
-        body = toJSON(wrapEntity(body = body)),
+        shojiURL(x, "catalogs", "scripts"),
+        body = toJSON(wrapEntity(body = list(body = script, ...))),
         status.handlers = list(`400` = crunchAutomationErrorHandler),
         progress.handler = crunchAutomationErrorHandler,
         config = add_headers(`Content-Type` = "application/json")
     )
-    # Provide feedback for dry_run success so that user is confident it was succesful
-    if (isTRUE(body$dry_run)) message("Script dry run was successful")
-    invisible(refresh(dataset))
-}
+
+    invisible(NULL)
+})
+
+setMethod("sendCrunchAutomationScript", "ProjectFolder", function(x,
+                                                                  script,
+                                                                  is_file = string_is_file_like(script),
+                                                                  encoding = "UTF-8",
+                                                                  ...) {
+    # project folders include a slot views with element execute,
+    # which gives us the URL to hit;
+    # but the account ('top-level folder', what you get from: `projects()`)
+    # is also of class ProjectFolder, but doesn't include this info;
+    # running CA scripts on the account is not supported currently
+    if (!is.crunchURL(x@views$execute)) {
+        halt(
+            "This folder does not support Crunch Automation scripts at this time."
+        )
+    }
+
+    dots <- list(...)
+    if (length(dots) > 0) {
+        # could have been a warning, but went with error in case a user
+        # would try running a destructive operation with dry_run = TRUE
+        stop("extra arguments (...) are not supported when x is a ProjectFolder")
+    }
+
+    crPOST(
+        shojiURL(x, "views", "execute"),
+        body = toJSON(wrapView(value = script)),
+        status.handlers = list(`400` = crunchAutomationErrorHandler),
+        progress.handler = crunchAutomationErrorHandler,
+        config = add_headers(`Content-Type` = "application/json")
+    )
+
+    invisible(NULL)
+})
 
 string_is_file_like <- function(x) {
     length(x) == 1 && # length 1 string
@@ -187,7 +250,9 @@ reset_automation_error_env <- function() {
 showScriptErrors <- function() {
     out <- as.list(automation_error_env)
 
-    if (is.null(out) || is.null(out$errors)) return(invisible(out))
+    if (is.null(out) || is.null(out$errors)) {
+        return(invisible(out))
+    }
 
     if (!is.null(out$file) && rstudio_markers_available()) {
         make_rstudio_markers(out)
@@ -225,7 +290,7 @@ crunchAutomationErrorHandler <- function(response) {
     if (inherits(response, "response")) {
         msg <- http_status(response)$message
         response_content <- content(response)
-    }  else {
+    } else {
         msg <- response$message$description
         response_content <- response$message
     }
@@ -258,7 +323,7 @@ crunchAutomationErrorHandler <- function(response) {
         # markers are available
         if (
             attr(error_items, "truncated") ||
-            (!is.null(automation_error_env$file) && rstudio_markers_available())
+                (!is.null(automation_error_env$file) && rstudio_markers_available())
         ) {
             more_info_text <- "\n\nRun command `showScriptErrors()` for more information."
         } else {
@@ -269,9 +334,9 @@ crunchAutomationErrorHandler <- function(response) {
     } else {
         # could also have information in message property
         # try to use it if it's a character string
-        other_msg <- try(content(response)[["message"]], silent = TRUE) #nocov
-        if (is.character(other_msg)) { #nocov
-            msg <- paste0(msg, " - ", paste0(other_msg, collapse = "\n")) #nocov
+        other_msg <- try(content(response)[["message"]], silent = TRUE) # nocov
+        if (is.character(other_msg)) { # nocov
+            msg <- paste0(msg, " - ", paste0(other_msg, collapse = "\n")) # nocov
         }
     }
     halt(msg)