+++ R/shoji.R
@@ -200,3 +200,7 @@ wrapEntity <- function(..., body = list(...)) {
 wrapCatalog <- function(...) list(element = "shoji:catalog", ...)
 
 wrapCatalogIndex <- function(...) wrapCatalog(index = as.list(...))
+
+wrapView <- function(value) {
+    list(element = "shoji:view", value = value)
+}